+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+- Rename `Subscription` to `TopicFilter`
 - Improve read-path error messages
 - Validate `SubscribePacket` / `UnsubscribePacket` topic filter list is not empty